Use a different config format
Instead of putting the metadata of each post in the post itself, all
post metadata is stored in a YAML file instead. The motivation behind
this is to make the Markdown documents less messy and more portable (as
they no longer contain PyMarkdown-specific metadata).

Also, I decided to drop TOML as a config format because I always forget
how it works :P